Add A* algorithm, mapFile error Fix
서버에 A* 알고리즘을 추가해서 Object의 움직임을 구현하였으며, 기존 mapFile을 만들 때 y좌표가 내려가 있는 것을 수정하였음.
기존 GameStart 로직을 전면 수정하고, 클라이언트에서 Unit을 선택해 움직이는 기능 추가.

+++ StarCraftClient/Assets/Script/ServerConnect.cs
@@ -121,7 +121,7 @@ public class ServerConnect : MonoBehaviour
     public Queue<Action> spawnQueue = new Queue<Action>();
 
     public uint currentUnitId;
-    public long playerIndex;
+    public int playerIndex;
 
     void Start()
     {
@@ -245,7 +245,7 @@ public class ServerConnect : MonoBehaviour
     }
 
 
-    public void EnqueueSpawn(UnitCode code, uint id, long owner, UnityEngine.Vector3 pos, UnityEngine.Quaternion dir, float hp, long spawnTime = 0)
+    public void EnqueueSpawn(UnitCode code, uint id, int owner, UnityEngine.Vector3 pos, UnityEngine.Quaternion dir, float hp, long spawnTime = 0)
     {
         spawnQueue.Enqueue(() => UnitManager.Instance.SpawnUnit(code, id, owner, pos, dir, hp, spawnTime));
     }
